requirements_dev.txt
aiofiles
aiohttp
aws-lambda-powertools
awslambdaric
bson
geopandas
jmespath
numpy
onnxruntime
opencv-python
pillow
pyproj
rasterio
requests
shapely
